requirements.txt
# ─── Web Framework ───────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ─── Database ─────────────────────────────────────────────────────────────────
motor==3.4.0
pymongo==4.7.2

# ─── Auth ─────────────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ─── Validation ───────────────────────────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1

# ─── File Parsing ─────────────────────────────────────────────────────────────
pdfplumber==0.11.0
python-docx==1.1.2
pypdf2==3.0.1

# ─── NLP & ML ─────────────────────────────────────────────────────────────────
scikit-learn==1.4.2
sentence-transformers==2.7.0
torch==2.2.2
transformers==4.40.1
numpy==1.26.4
nltk==3.8.1
spacy==3.7.4

# ─── PDF Generation ───────────────────────────────────────────────────────────
reportlab==4.1.0
weasyprint==62.1
jinja2==3.1.4

# ─── HTTP Client (GitHub API) ─────────────────────────────────────────────────
httpx==0.27.0
aiohttp==3.9.5

# ─── Background Tasks ─────────────────────────────────────────────────────────
celery==5.4.0
redis==5.0.4

# ─── Config & Env ─────────────────────────────────────────────────────────────
python-dotenv==1.0.1

# ─── Logging & Monitoring ─────────────────────────────────────────────────────
structlog==24.1.0
prometheus-client==0.20.0
sentry-sdk[fastapi]==2.1.1

# ─── Caching ──────────────────────────────────────────────────────────────────
aiocache==0.12.2

# ─── Testing ──────────────────────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
faker==25.0.0

# ─── Utils ────────────────────────────────────────────────────────────────────
python-dateutil==2.9.0
chardet==5.2.0
pillow==10.3.0